backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.0
lxml==5.3.0
openai==1.57.0
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.19
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.32.1
